+++ js/main.js
@@ -2,3 +2,4 @@ import './util.js';
 import './data.js';
 import './popup.js';
 import './form.js';
+import './map.js';
